+++ src/main/java/com/ffuntree/ffunfun/exception/AuthorizationExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.ffuntree.ffunfun.exception;
 
-import com.ffuntree.ffunfun.data.dto.ErrorResponseDto;
+import com.ffuntree.ffunfun.data.common.ErrorResponseDto;
 import io.jsonwebtoken.JwtException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.AuthorizationServiceException;
